+++ angular-workspace/projects/ni/nimble-angular/src/directives/table-column/text/tests/nimble-table-column-text.directive.spec.ts
@@ -30,6 +30,8 @@ describe('NimbleTableColumnText', () => {
                         action-menu-slot="my-slot"
                         action-menu-label="my menu"
                         column-hidden="true"
+                        fractional-width="2"
+                        min-pixel-width="40"
                         sort-direction="${TableColumnSortDirection.ascending}"
                         sort-index="0"
                     ></nimble-table-column-text>
@@ -95,6 +97,21 @@ describe('NimbleTableColumnText', () => {
             expect(directive.sortIndex).toBe(0);
             expect(nativeElement.sortIndex).toBe(0);
         });
+
+        it('will use template string values for fractionalWidth', () => {
+            expect(directive.fractionalWidth).toBe(2);
+            expect(nativeElement.fractionalWidth).toBe(2);
+        });
+
+        it('will use template string values for minPixelWidth', () => {
+            expect(directive.minPixelWidth).toBe(40);
+            expect(nativeElement.minPixelWidth).toBe(40);
+        });
+
+        it('will use template string values for minPixelWidth', () => {
+            expect(directive.minPixelWidth).toBe(40);
+            expect(nativeElement.minPixelWidth).toBe(40);
+        });
     });
 
     describe('with property bound values', () => {
@@ -109,6 +126,8 @@ describe('NimbleTableColumnText', () => {
                         [actionMenuSlot]="actionMenuSlot"
                         [actionMenuLabel]="actionMenuLabel"
                         [column-hidden]="columnHidden"
+                        [fractional-width]="fractionalWidth"
+                        [min-pixel-width]="minPixelWidth"
                         [sort-direction]="sortDirection"
                         [sort-index]="sortIndex"
                     ></nimble-table-column-text>
@@ -122,6 +141,8 @@ describe('NimbleTableColumnText', () => {
             public placeholder = 'no value';
             public actionMenuSlot = 'my-slot';
             public actionMenuLabel = 'my menu';
+            public fractionalWidth: number | null = 2;
+            public minPixelWidth: number | null = 40;
             public columnId = 'my-column';
             public columnHidden = true;
             public sortDirection: TableColumnSortDirection = TableColumnSortDirection.ascending;
@@ -241,6 +262,50 @@ describe('NimbleTableColumnText', () => {
             expect(directive.sortIndex).toBe(null);
             expect(nativeElement.sortIndex).toBe(null);
         });
+
+        it('can be configured with property binding for fractionalWidth', () => {
+            expect(directive.fractionalWidth).toBe(2);
+            expect(nativeElement.fractionalWidth).toBe(2);
+
+            fixture.componentInstance.fractionalWidth = 1;
+            fixture.detectChanges();
+
+            expect(directive.fractionalWidth).toBe(1);
+            expect(nativeElement.fractionalWidth).toBe(1);
+        });
+
+        it('can be configured with property binding for fractionalWidth updated to null', () => {
+            expect(directive.fractionalWidth).toBe(2);
+            expect(nativeElement.fractionalWidth).toBe(2);
+
+            fixture.componentInstance.fractionalWidth = null;
+            fixture.detectChanges();
+
+            expect(directive.fractionalWidth).toBe(null);
+            expect(nativeElement.fractionalWidth).toBe(null);
+        });
+
+        it('can be configured with property binding for minPixelWidth', () => {
+            expect(directive.minPixelWidth).toBe(40);
+            expect(nativeElement.minPixelWidth).toBe(40);
+
+            fixture.componentInstance.minPixelWidth = 50;
+            fixture.detectChanges();
+
+            expect(directive.minPixelWidth).toBe(50);
+            expect(nativeElement.minPixelWidth).toBe(50);
+        });
+
+        it('can be configured with property binding for minPixelWidth updated to null', () => {
+            expect(directive.minPixelWidth).toBe(40);
+            expect(nativeElement.minPixelWidth).toBe(40);
+
+            fixture.componentInstance.minPixelWidth = null;
+            fixture.detectChanges();
+
+            expect(directive.minPixelWidth).toBe(null);
+            expect(nativeElement.minPixelWidth).toBe(null);
+        });
     });
 
     describe('with attribute bound values', () => {
@@ -255,6 +320,8 @@ describe('NimbleTableColumnText', () => {
                         [attr.action-menu-slot]="actionMenuSlot"
                         [attr.action-menu-label]="actionMenuLabel"
                         [attr.column-hidden]="columnHidden"
+                        [attr.fractional-width]="fractionalWidth"
+                        [attr.min-pixel-width]="minPixelWidth"
                         [attr.sort-direction]="sortDirection"
                         [attr.sort-index]="sortIndex"
                     ></nimble-table-column-text>
@@ -268,6 +335,8 @@ describe('NimbleTableColumnText', () => {
             public placeholder = 'no value';
             public actionMenuSlot = 'my-slot';
             public actionMenuLabel = 'my menu';
+            public fractionalWidth: number | null = 2;
+            public minPixelWidth: number | null = 40;
             public columnId = 'my-column';
             public columnHidden = true;
             public sortDirection: TableColumnSortDirection = TableColumnSortDirection.ascending;
@@ -386,6 +455,50 @@ describe('NimbleTableColumnText', () => {
 
             expect(directive.sortIndex).toBe(null);
             expect(nativeElement.sortIndex).toBe(null);
+        });
+
+        it('can be configured with attribute binding for fractionalWidth', () => {
+            expect(directive.fractionalWidth).toBe(2);
+            expect(nativeElement.fractionalWidth).toBe(2);
+
+            fixture.componentInstance.fractionalWidth = 1;
+            fixture.detectChanges();
+
+            expect(directive.fractionalWidth).toBe(1);
+            expect(nativeElement.fractionalWidth).toBe(1);
+        });
+
+        it('can be configured with attribute binding for fractionalWidth set to null', () => {
+            expect(directive.fractionalWidth).toBe(2);
+            expect(nativeElement.fractionalWidth).toBe(2);
+
+            fixture.componentInstance.fractionalWidth = null;
+            fixture.detectChanges();
+
+            expect(directive.fractionalWidth).toBe(null);
+            expect(nativeElement.fractionalWidth).toBe(null);
+        });
+
+        it('can be configured with attribute binding for minPixelWidth', () => {
+            expect(directive.minPixelWidth).toBe(40);
+            expect(nativeElement.minPixelWidth).toBe(40);
+
+            fixture.componentInstance.minPixelWidth = 50;
+            fixture.detectChanges();
+
+            expect(directive.minPixelWidth).toBe(50);
+            expect(nativeElement.minPixelWidth).toBe(50);
+        });
+
+        it('can be configured with attribute binding for minPixelWidth set to null', () => {
+            expect(directive.minPixelWidth).toBe(40);
+            expect(nativeElement.minPixelWidth).toBe(40);
+
+            fixture.componentInstance.minPixelWidth = null;
+            fixture.detectChanges();
+
+            expect(directive.minPixelWidth).toBe(null);
+            expect(nativeElement.minPixelWidth).toBe(null);
         });
     });
 });