+++ packages/nimble-components/src/table/components/cell/index.ts
@@ -63,3 +63,4 @@ const nimbleTableCell = TableCell.compose({
 });
 
 DesignSystem.getOrCreate().withPrefix('nimble').register(nimbleTableCell());
+export const tableCellTag = DesignSystem.tagFor(TableCell);