+++ packages/nimble-components/src/table/components/row/index.ts
@@ -124,3 +124,4 @@ const nimbleTableRow = TableRow.compose({
 });
 
 DesignSystem.getOrCreate().withPrefix('nimble').register(nimbleTableRow());
+export const tableRowTag = DesignSystem.tagFor(TableRow);