+++ packages/nimble-components/src/table/components/row/styles.ts
@@ -10,10 +10,13 @@ export const styles = css`
     ${display('grid')}
 
     :host {
-        height: ${controlHeight};
+        height: calc(${controlHeight} + 2 * ${borderWidth});
         border-top: calc(2 * ${borderWidth}) solid ${tableRowBorderColor};
         grid-auto-flow: column;
         grid-auto-columns: 1fr;
+        grid-template-columns: var(--ni-private-table-row-grid-columns) auto;
+        width: fit-content;
+        min-width: 100%;
     }
 
     nimble-table-cell {